지하철 OpenAPI 활용 DB 데이터 접근

+++ src/main/java/com/busanit/busan_subway_project/controller/LineController.java
@@ -14,7 +14,7 @@ public class LineController {
     private LineService lineService;
 
     @GetMapping("/line/{line_cd}")
-    public Line getLineByCode(@PathVariable("line_cd") Long line_cd){
+    public Line getLineByCode(@PathVariable("line_cd") int line_cd){
         return lineService.getLine(line_cd);
     }
 }
