+++ src/main/java/com/busanit/busan_subway_project/service/StationLocationService.java
@@ -11,7 +11,7 @@ public class StationLocationService {
     private StationLocationRepo stationLocationRepo;
 
     // Station으로 x1, y1, x2, y2 출력
-    public Station_Location getStationLocation(Long scode){
+    public Station_Location getStationLocation(int scode){
         Station_Location sl = stationLocationRepo.findStationLocationByScode(scode);
         return new Station_Location(sl.getScode(), sl.getStation(), sl.getX1(), sl.getY1(), sl.getX2(), sl.getY2());
     }
